requirements.txt
requests
bs4
python-telegram-bot